Random now returns 64bit integers, Transposition table stores moves for both sides now

+++ Engine.Tests.Unit/StubRng.cpp
@@ -2,7 +2,7 @@
 
 namespace engine {
 
-	StubRng::StubRng(int values[]) {
+	StubRng::StubRng(long long values[]) {
 		_pointer = 0;
 		_values = values;
 	}
@@ -11,7 +11,7 @@ namespace engine {
 		delete _values;
 	}
 
-	int StubRng::NextInteger() {
+	long long StubRng::NextInteger() {
 		return _values[_pointer++];
 	}
 }
